[add] バグまみれだけどいったんプッシュ
