+++ common/api/env_sample.js
@@ -1,3 +1,2 @@
-export const GAS_URL = "";
 export const NOTION_API_TOKEN = "";
 export const NOTION_DATABASE_ID = "";
