+++ manifest.json
@@ -14,7 +14,10 @@
   ],
   "web_accessible_resources": [
     {
-      "resources": ["content_scripts/content.js"],
+      "resources": [
+        "content_scripts/content.js",
+        "content_scripts/parser/parser.js"
+      ],
       "matches": ["https://atcoder.jp/*"]
     }
   ],
